fixers/api-breaking-change-detect/requirements.txt
langchain
langchain-openai
openai
gitpython
rich
typer
pyyaml
pytest
pydantic
python-dotenv
